pyproject.toml
[project]
name = "api-electrolux"
version = "0.1.0"
description = ""
authors = [
    {name = " Isaias-Felix",email = "desenvolvimento1@novalar.com.br"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "requests (>=2.32.3,<3.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
dev = 'fastapi dev api_electrolux/main.py'
run = 'fastapi run api_electrolux/main.py'